chore: fix spelling mistakes and rename watering_plan to wateringplan
