+++ internal/entities/watering_plan.go
@@ -9,7 +9,7 @@ type WateringPlanStatus string
 const (
 	WateringPlanStatusPlanned     WateringPlanStatus = "planned"
 	WateringPlanStatusActive      WateringPlanStatus = "active"
-	WateringPlanStatusCancelled   WateringPlanStatus = "cancelled"
+	WateringPlanStatusCanceled   WateringPlanStatus = "canceled"
 	WateringPlanStatusFinished    WateringPlanStatus = "finished"
 	WateringPlanStatusNotCompeted WateringPlanStatus = "not competed"
 	WateringPlanStatusUnknown     WateringPlanStatus = "unknown"
