+++ internal/storage/postgres/wateringplan/create_test.go
@@ -1,4 +1,4 @@
-package watering_plan
+package wateringplan
 
 import (
 	"context"
