+++ internal/storage/postgres/wateringplan/update_test.go
@@ -1,4 +1,4 @@
-package watering_plan
+package wateringplan
 
 import (
 	"context"
