+++ internal/storage/postgres/wateringplan/watering_plan.go
@@ -1,4 +1,4 @@
-package watering_plan
+package wateringplan
 
 import (
 	"context"
